+++ crates/nodes/stereo_visual_odometry/Cargo.toml
@@ -12,6 +12,7 @@ nalgebra = { workspace = true }
 projection = { workspace = true }
 ros-z = { workspace = true }
 ros2 = { workspace = true }
+serde = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "signal"] }
 types = { workspace = true }
 ort = { workspace = true, features = ["cuda", "tensorrt"] }
