+++ crates/types/src/parameters.rs
@@ -673,73 +673,6 @@ pub struct DetectionParameters {
     PathIntrospect,
     ros_z::Message,
 )]
-pub struct StereoVisualOdometryParameters {
-    pub enable: bool,
-    pub neural_networks_folder: PathBuf,
-    pub model_name: String,
-    pub pose_estimation_parameters: StereoVisualOdometryPoseEstimationParameters,
-}
-
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Deserialize,
-    Serialize,
-    PathSerialize,
-    PathDeserialize,
-    PathIntrospect,
-    ros_z::Message,
-)]
-pub struct StereoVisualOdometryPoseEstimationParameters {
-    pub minimum_pnp_correspondences: usize,
-    pub ransac_reprojection_threshold_px: f32,
-    pub ransac_max_iterations: usize,
-    pub ransac_confidence: f32,
-    pub lm_max_iterations: usize,
-    pub lm_initial_lambda: f32,
-    pub lm_min_lambda: f32,
-    pub lm_max_lambda: f32,
-    pub lm_step_tolerance: f32,
-    pub lm_cost_tolerance: f32,
-    pub lm_huber_threshold_px: f32,
-    pub full_weight_disparity_px: f32,
-    pub min_disparity_weight: f32,
-    pub max_vertical_disparity_px: f32,
-}
-
-impl Default for StereoVisualOdometryPoseEstimationParameters {
-    fn default() -> Self {
-        Self {
-            minimum_pnp_correspondences: 8,
-            ransac_reprojection_threshold_px: 6.0,
-            ransac_max_iterations: 100,
-            ransac_confidence: 0.99,
-            lm_max_iterations: 20,
-            lm_initial_lambda: 1e-3,
-            lm_min_lambda: 1e-7,
-            lm_max_lambda: 1e10,
-            lm_step_tolerance: 1e-6,
-            lm_cost_tolerance: 1e-6,
-            lm_huber_threshold_px: 3.0,
-            full_weight_disparity_px: 8.0,
-            min_disparity_weight: 0.5,
-            max_vertical_disparity_px: 3.0,
-        }
-    }
-}
-
-#[derive(
-    Clone,
-    Debug,
-    Default,
-    Deserialize,
-    Serialize,
-    PathSerialize,
-    PathDeserialize,
-    PathIntrospect,
-    ros_z::Message,
-)]
 pub struct ObjectDetectionParameters {
     pub maximum_intersection_over_union: f32,
     pub confidence_threshold: f32,
